+++ ui/src/pages/signin/SignIn.css
@@ -1,0 +1,3 @@
+.errorMessage {
+  color: red;
+}
